`config` may not exist
In my repo. [ckeditor5-build-full](https://github.com/Eusen/ckeditor5-build-full)
I've mixed five kinds of editors together to distinguish which one should be created by the type of create.
Because in a real project, it is very common to mix multiple editors.

But there is a problem here when using: `this.editor.config.get( 'blockToolbar' )` may not exist.
So we need to add a judgment here. If it doesn't exist, we won't deal with it.